CPickup(s) further (#494)

+++ source/game_sa/Entity/Object/Object.cpp
@@ -173,7 +173,7 @@ void* CObject::operator new(size_t size) {
 }
 
 // 0x5A1EF0
-void* CObject::operator new(unsigned size, int32 poolRef) {
+void* CObject::operator new(size_t size, int32 poolRef) {
     return GetObjectPool()->NewAt(poolRef);
 }
